Outsourcing of the time cut parameters (and others)
Attempted to create several functions to that the time cuts applied e.g. to the NaI spectra are read from the batchfile instead of the user_sort.

Additionally, the PPAC channels and some other variables appear in there now.

After the data is read, it's given some name (in a struc) and usually processed such, that one can later call a function like "CheckNaIpromtGate (float)" with a float like the time nai_t to check whether nai_t is in the prompt Gate of the sodium iodines

TODO:
Make the program work also when the define USE_FISSION_PARAMETERS is set to zero. For example a section labels with TODO needs to be modified for this.
(Problem: I still use IsPPACChennel in some if conditions to fill spectra, however the PPAC Channel definition only makes sense if we used them).