Fix of a nasty bug!
Previously we tried to fill ppac matrix indexed by the e-detector (variable 'ei') this obviously failed because we only have 4 PPACS, thus causing a segmentation fault. This was previously fixed by allocating 8 m_ppac_e_t matrices (see line 361). However, m_ppac_e_t is an array of size 4, thus allocating stuff should beyond index 3 should have given a segmentation fault or similar. However it turns out that for some versions of gcc/clang the padding of the UserXY object is such in memory that we are indeed able to allocate 8 m_ppac_e_t matrices without any errors and the sorting routine did not fail. With newer versions of GCC it seems like the packing of the object in memory have changed slightly and we will somehow overwrite something else and cause the computer to crash.