Sorting routine GammasOnly, no conincidence
Copied out the sorting routine for gammas only, eg. if one wants to sort spectra of radioactive sources.
I left most parts of the parameter input unchanged, so that one could simply use the same "batch file" and "gainshifts file" as for other runs. This means, however, that many unused parameters will be read in.